refactor(showing test): refactor functions of showing test and cconstants
