+++ src/modules/showing-test/ShowingTest.tsx
@@ -1,10 +1,18 @@
 import { ReactNode, useEffect, useState } from 'react';
-import nlp from 'compromise';
 import { Button } from '@/ui-components/Button';
 import { Icon } from '@/ui-components/Icon';
 import { CircleButton } from '@/ui-components/CircleButton';
 import { DEFAULT_TEST } from '@/constants/data';
 import styles from './showingTest.module.css';
+import {
+	ARTICLES,
+	PRONOUN_CATEGORIES,
+	CONJUCTIONS_CATEGORIES,
+	PREPOSITION_CATEGORIES,
+	MAX_VARIANTS,
+} from './constants';
+import { detectPartOfSpeech } from './functions/detectPartOfSpeech';
+import { showNounsOrAdverbs, showVariants, showVerbs, getGroup } from './functions/showVariants';
 
 type Test = {
 	id: number;
@@ -19,159 +27,7 @@ type Answer = {
 	isCorrect: boolean;
 };
 
-type NounData = {
-	word: string;
-	score: number;
-};
-
-type VerbList = {
-	[key: string]: string;
-};
-
-const pronounCategories = {
-	subject: ['I', 'you', 'he', 'she', 'it', 'we', 'they'],
-	object: ['me', 'you', 'him', 'her', 'it', 'us', 'them'],
-	indefinite: ['some', 'any', 'no', 'little', 'few', 'many', 'much'],
-	interrogative: ['who', 'what', 'which', 'whose', 'whom'],
-	relative: ['who', 'what', 'which', 'whose', 'whom'],
-	possessive: ['my', 'your', 'his', 'her', 'its', 'our', 'their'],
-	reflexive: ['myself', 'yourself', 'himself', 'herself', 'itself', 'ourselves', 'themselves'],
-};
-
-const prepositionCategories = {
-	simple: [
-		'about',
-		'above',
-		'across',
-		'after',
-		'against',
-		'along',
-		'among',
-		'around',
-		'at',
-		'before',
-		'behind',
-		'below',
-		'beneath',
-		'beside',
-		'between',
-		'beyond',
-		'by',
-		'down',
-		'during',
-		'except',
-		'for',
-		'from',
-		'in',
-		'into',
-		'near',
-		'of',
-		'off',
-		'on',
-		'out',
-		'over',
-		'past',
-		'since',
-		'through',
-		'to',
-		'toward',
-		'under',
-		'until',
-		'up',
-		'with',
-		'within',
-		'without',
-	],
-	compound: [
-		'according to',
-		'ahead of',
-		'apart from',
-		'as for',
-		'as per',
-		'as to',
-		'because of',
-		'by means of',
-		'in accordance with',
-		'in addition to',
-		'in case of',
-		'in front of',
-		'in lieu of',
-		'in place of',
-		'in spite of',
-		'instead of',
-		'on account of',
-		'on behalf of',
-		'on top of',
-		'out of',
-		'prior to',
-		'with regard to',
-		'with respect to',
-	],
-	phrasal: [
-		'along with',
-		'away from',
-		'close to',
-		'due to',
-		'except for',
-		'far from',
-		'next to',
-		'owing to',
-		'rather than',
-		'thanks to',
-	],
-	place: ['on', 'in', 'under', 'between', 'next to'],
-	direction: ['to', 'from', 'out of', 'into', 'onto', 'towards'],
-	time: ['at', 'in', 'before', 'after', 'by', 'during'],
-	reasons: ['because of', 'due to', 'for', 'to', 'in order to'],
-};
-
-const conjustionsCategories = {
-	coordinating: ['for', 'and', 'nor', 'but', 'or', 'yet', 'so'],
-	time: [
-		'after',
-		'as',
-		'as soon as',
-		'before',
-		'once',
-		'since',
-		'till',
-		'until',
-		'when',
-		'whenever',
-		'while',
-	],
-	reason: ['because', 'as', 'since', 'so that', 'in order that'],
-	condition: ['if', 'unless', 'even if', 'provided that'],
-	contrast: ['although', 'though', 'even though', 'whereas', 'while'],
-	comparison: ['as', 'than'],
-	conjuctive: [
-		'accordingly',
-		'also',
-		'besides',
-		'consequently',
-		'finally',
-		'furthermore',
-		'hence',
-		'however',
-		'indeed',
-		'instead',
-		'likewise',
-		'moreover',
-		'nevertheless',
-		'nonetheless',
-		'otherwise',
-		'similarly',
-		'still',
-		'then',
-		'therefore',
-		'thus',
-	],
-};
-
-const articles = ['a', 'an', 'the'];
-
 export const ShowingTest = () => {
-	const maxNumbersOfAnswers = 8;
 	const [answers, setAnswers] = useState<Answer[]>([]);
 	const [chooseAnswer, setChooseAnswer] = useState<string>('');
 
@@ -188,134 +44,38 @@ export const ShowingTest = () => {
 	const [answeredQuestions, setAnsweredQuestions] = useState(0);
 
 	const fetchDefinition = async (word: string) => {
-		const doc = nlp(word);
-		const isVerb = doc.verbs().out('array').length > 0;
-		const isPronouns = doc.pronouns().out('array').length > 0;
-		const isNoun = doc.nouns().out('array').length > 0;
-		const isPreposition = doc.prepositions().out('array').length > 0;
-		const isArticle = articles.includes(word.toLocaleLowerCase());
-		const isAdjective = doc.adjectives().out('array').length > 0;
-		const isConjuction = doc.conjunctions().out('array').length > 0;
+		const { isAdjective, isArticle, isConjuction, isNoun, isPreposition, isVerb, isPronoun } =
+			detectPartOfSpeech(word);
 
 		if (isVerb) {
-			const verbList = doc.verbs().conjugate()[0] as VerbList;
-			const answers = Object.keys(verbList)
-				.map((key) => verbList[key])
-				.map((item, index) => {
-					if (item.toLowerCase() === word.toLowerCase()) {
-						return { id: index + 1, name: item, isCorrect: true };
-					} else {
-						return { id: index + 1, name: item, isCorrect: false };
-					}
-				});
-			setAnswers(shuffle([...answers]));
-		} else if (
-			(isNoun &&
-				!pronounCategories.object.includes(word) &&
-				!pronounCategories.possessive.includes(word) &&
-				!pronounCategories.reflexive.includes(word) &&
-				!pronounCategories.subject.includes(word) &&
-				!pronounCategories.relative.includes(word) &&
-				!pronounCategories.interrogative.includes(word)) ||
-			isAdjective
-		) {
-			const response = await fetch(`https://api.datamuse.com/words?rel_syn=${word}`);
-			const data = await response.json();
-			const dataArray = data.map((item: NounData) => item.word);
-			const result = makeMaxAllowedAnswers(dataArray, word);
-
-			setAnswers(shuffle([...result]));
-		} else if (isPronouns) {
-			if (pronounCategories.object.includes(word)) {
-				const answers = getGroup(pronounCategories.object, word);
-				setAnswers(shuffle([...answers]));
-			}
-
-			if (pronounCategories.possessive.includes(word)) {
-				const answers = getGroup(pronounCategories.possessive, word);
-				setAnswers(shuffle([...answers]));
-			}
-
-			if (pronounCategories.reflexive.includes(word)) {
-				const answers = getGroup(pronounCategories.reflexive, word);
-				setAnswers(shuffle([...answers]));
-			}
-
-			if (pronounCategories.subject.includes(word)) {
-				const answers = getGroup(pronounCategories.subject, word);
-				setAnswers(shuffle([...answers]));
-			}
-
-			if (pronounCategories.relative.includes(word)) {
-				const answers = getGroup(pronounCategories.relative, word);
-				setAnswers(shuffle([...answers]));
-			}
-
-			if (pronounCategories.interrogative.includes(word)) {
-				const answers = getGroup(pronounCategories.interrogative, word);
-				setAnswers(shuffle([...answers]));
-			}
-		} else if (isPreposition) {
-			if (prepositionCategories.direction.includes(word)) {
-				const answers = makeMaxAllowedAnswers(prepositionCategories.direction, word);
-				setAnswers(shuffle([...answers]));
-			}
-
-			if (prepositionCategories.place.includes(word)) {
-				const answers = getGroup(prepositionCategories.place, word);
-				setAnswers(shuffle([...answers]));
-			}
-
-			if (prepositionCategories.reasons.includes(word)) {
-				const answers = makeMaxAllowedAnswers(prepositionCategories.reasons, word);
-				setAnswers(shuffle([...answers]));
-			}
-
-			if (prepositionCategories.time.includes(word)) {
-				const answers = makeMaxAllowedAnswers(prepositionCategories.time, word);
-				setAnswers(shuffle([...answers]));
-			}
-
-			if (prepositionCategories.simple.includes(word)) {
-				const result = makeMaxAllowedAnswers(prepositionCategories.simple, word);
-				setAnswers(shuffle([...result]));
-			}
-		} else if (isArticle) {
-			const answers = getGroup(articles, word);
-			setAnswers(shuffle([...answers]));
-		} else if (isConjuction) {
-			if (conjustionsCategories.contrast) {
-				const answers = getGroup(conjustionsCategories.contrast, word);
-				setAnswers(shuffle([...answers]));
-			} else if (conjustionsCategories.coordinating.includes(word)) {
-				const answers = getGroup(conjustionsCategories.coordinating, word);
-				setAnswers(shuffle([...answers]));
-			} else if (conjustionsCategories.conjuctive.includes(word)) {
-				const answers = getGroup(conjustionsCategories.conjuctive, word);
-				setAnswers(shuffle([...answers]));
-			} else if (conjustionsCategories.comparison) {
-				const answers = getGroup(conjustionsCategories.comparison, word);
-				setAnswers(shuffle([...answers]));
-			} else if (conjustionsCategories.condition) {
-				const answers = getGroup(conjustionsCategories.condition, word);
-				setAnswers(shuffle([...answers]));
-			} else if (conjustionsCategories.reason) {
-				const answers = getGroup(conjustionsCategories.reason, word);
-				setAnswers(shuffle([...answers]));
-			} else if (conjustionsCategories.time) {
-				const answers = getGroup(conjustionsCategories.time, word);
-				setAnswers(shuffle([...answers]));
-			}
+			const answers = showVerbs(word);
+			setAnswers([...answers]);
 		}
-	};
 
-	const shuffle = (array: Answer[]) => {
-		const newArray = [...array];
-		for (let i = newArray.length - 1; i > 0; i--) {
-			const j = Math.floor(Math.random() * (i + 1));
-			[newArray[i], newArray[j]] = [newArray[j], newArray[i]];
+		if ((isNoun && !isPronoun) || isAdjective) {
+			const result = await showNounsOrAdverbs(word, MAX_VARIANTS);
+			setAnswers([...result]);
 		}
-		return newArray;
+
+		if (isPronoun) {
+			const answers = showVariants(PRONOUN_CATEGORIES, word);
+			setAnswers([...answers]);
+		}
+
+		if (isPreposition) {
+			const answers = showVariants(PREPOSITION_CATEGORIES, word);
+			setAnswers([...answers]);
+		}
+
+		if (isArticle) {
+			const answers = getGroup(ARTICLES, word.toLocaleLowerCase());
+			setAnswers([...answers]);
+		}
+
+		if (isConjuction) {
+			const answers = showVariants(CONJUCTIONS_CATEGORIES, word);
+			setAnswers([...answers]);
+		}
 	};
 
 	const onClick = (answer: Answer) => {
@@ -347,41 +107,6 @@ export const ShowingTest = () => {
 		}
 
 		setChooseAnswer(chooseAnswer + ` ${answer.name}`);
-	};
-
-	const getGroup = (variants: string[], word: string) => {
-		const answers = variants.map((item, index) => {
-			if (item.toLocaleLowerCase() === word.toLocaleLowerCase()) {
-				return { id: index + 1, name: item, isCorrect: true };
-			} else {
-				return { id: index + 1, name: item, isCorrect: false };
-			}
-		});
-
-		return answers;
-	};
-
-	const makeMaxAllowedAnswers = (variants: string[], correctAnswer: string) => {
-		let result: Answer[] = [];
-		for (let i = 0; i <= maxNumbersOfAnswers - 1; i++) {
-			if (i === maxNumbersOfAnswers - 1) {
-				result = [
-					...result,
-					{ id: i + 1, name: correctAnswer.toLocaleLowerCase(), isCorrect: true },
-				];
-			} else {
-				result = [
-					...result,
-					{
-						id: i + 1,
-						name: variants[i],
-						isCorrect: false,
-					},
-				];
-			}
-		}
-
-		return result;
 	};
 
 	const swapQuestion = (id: number) => {
