+++ src/modules/showing-test/constants.ts
@@ -1,0 +1,145 @@
+import type { Variants } from './functions/showVariants';
+
+export const MAX_VARIANTS = 8;
+
+export const PRONOUN_CATEGORIES: Variants = {
+	subject: ['I', 'you', 'he', 'she', 'it', 'we', 'they'],
+	object: ['me', 'you', 'him', 'her', 'it', 'us', 'them'],
+	indefinite: ['some', 'any', 'no', 'little', 'few', 'many', 'much'],
+	interrogative: ['who', 'what', 'which', 'whose', 'whom'],
+	relative: ['who', 'what', 'which', 'whose', 'whom'],
+	possessive: ['my', 'your', 'his', 'her', 'its', 'our', 'their'],
+	reflexive: ['myself', 'yourself', 'himself', 'herself', 'itself', 'ourselves', 'themselves'],
+};
+
+export const PREPOSITION_CATEGORIES: Variants = {
+	simple: [
+		'about',
+		'above',
+		'across',
+		'after',
+		'against',
+		'along',
+		'among',
+		'around',
+		'at',
+		'before',
+		'behind',
+		'below',
+		'beneath',
+		'beside',
+		'between',
+		'beyond',
+		'by',
+		'down',
+		'during',
+		'except',
+		'for',
+		'from',
+		'in',
+		'into',
+		'near',
+		'of',
+		'off',
+		'on',
+		'out',
+		'over',
+		'past',
+		'since',
+		'through',
+		'to',
+		'toward',
+		'under',
+		'until',
+		'up',
+		'with',
+		'within',
+		'without',
+	],
+	compound: [
+		'according to',
+		'ahead of',
+		'apart from',
+		'as for',
+		'as per',
+		'as to',
+		'because of',
+		'by means of',
+		'in accordance with',
+		'in addition to',
+		'in case of',
+		'in front of',
+		'in lieu of',
+		'in place of',
+		'in spite of',
+		'instead of',
+		'on account of',
+		'on behalf of',
+		'on top of',
+		'out of',
+		'prior to',
+		'with regard to',
+		'with respect to',
+	],
+	phrasal: [
+		'along with',
+		'away from',
+		'close to',
+		'due to',
+		'except for',
+		'far from',
+		'next to',
+		'owing to',
+		'rather than',
+		'thanks to',
+	],
+	place: ['on', 'in', 'under', 'between', 'next to'],
+	direction: ['to', 'from', 'out of', 'into', 'onto', 'towards'],
+	time: ['at', 'in', 'before', 'after', 'by', 'during'],
+	reasons: ['because of', 'due to', 'for', 'to', 'in order to'],
+};
+
+export const CONJUCTIONS_CATEGORIES: Variants = {
+	coordinating: ['for', 'and', 'nor', 'but', 'or', 'yet', 'so'],
+	time: [
+		'after',
+		'as',
+		'as soon as',
+		'before',
+		'once',
+		'since',
+		'till',
+		'until',
+		'when',
+		'whenever',
+		'while',
+	],
+	reason: ['because', 'as', 'since', 'so that', 'in order that'],
+	condition: ['if', 'unless', 'even if', 'provided that'],
+	contrast: ['although', 'though', 'even though', 'whereas', 'while'],
+	comparison: ['as', 'than'],
+	conjuctive: [
+		'accordingly',
+		'also',
+		'besides',
+		'consequently',
+		'finally',
+		'furthermore',
+		'hence',
+		'however',
+		'indeed',
+		'instead',
+		'likewise',
+		'moreover',
+		'nevertheless',
+		'nonetheless',
+		'otherwise',
+		'similarly',
+		'still',
+		'then',
+		'therefore',
+		'thus',
+	],
+};
+
+export const ARTICLES: string[] = ['a', 'an', 'the'];
